distinct data module for calendar feature
