+++ app/src/main/java/org/stepik/android/view/injection/course/CourseComponent.kt
@@ -7,6 +7,7 @@ import org.stepik.android.view.course.ui.activity.CourseActivity
 import org.stepik.android.view.course_content.ui.fragment.CourseContentFragment
 import org.stepik.android.view.course_info.ui.fragment.CourseInfoFragment
 import org.stepik.android.view.injection.billing.BillingDataModule
+import org.stepik.android.view.injection.calendar.CalendarDataModule
 import org.stepik.android.view.injection.course_calendar.CourseCalendarDataModule
 import org.stepik.android.view.injection.course_list.CourseListDataModule
 import org.stepik.android.view.injection.course_payments.CoursePaymentsDataModule
@@ -32,7 +33,8 @@ import org.stepik.android.view.injection.user.UserDataModule
 
     BillingDataModule::class,
 
-    CourseCalendarDataModule::class
+    CourseCalendarDataModule::class,
+    CalendarDataModule::class
 ])
 interface CourseComponent {
     @Subcomponent.Builder
